Moves temporal interp creation to fluid.py
